chore(create-email): move CLI code into the command away from createStarter
